LIBITD-969 Add validation to ensure unit is in dept
This adds a validation to check if the request's unit is in the department
Also moves the Request class methods to a concern, since Rubocop
complained that the class was too long

https://issues.umd.edu/browse/LIBITD-969
